requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
python-dotenv==1.0.1
email_validator==2.1.1
